requirements.txt
Django>=4.2.0
Pillow>=10.0.0
python-telegram-bot>=20.0
requests>=2.31.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
django-allauth>=0.54.0
django-ckeditor>=6.5.0
django-import-export>=3.2.0
django-redis>=5.4.0
django-cors-headers>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0